split: interval_test.py to test/interval_methods
